created separate utility and topic options file
